+++ src/Task/Git.php
@@ -6,12 +6,9 @@ use Robo\Result;
 use Robo\Task\Shared\CommandInterface;
 use Robo\Task\Shared\TaskInterface;
 
-/**
- * BundledTasks to do Git stuff
- */
 trait Git {
 
-    protected function taskGit($pathToGit = 'git')
+    protected function taskGitStack($pathToGit = 'git')
     {
         return new GitStackTask($pathToGit);
     }
@@ -23,14 +20,14 @@ trait Git {
  *
  * ``` php
  * <?php
- * $this->taskGit()
+ * $this->taskGitStack()
  *  ->stopOnFail()
  *  ->add('-A')
  *  ->commit('adding everything')
  *  ->push('origin','master')
  *  ->run()
  *
- * $this->taskGit()
+ * $this->taskGitStack()
  *  ->stopOnFail()
  *  ->add('doc/*')
  *  ->commit('doc updated')
@@ -55,42 +52,87 @@ class GitStackTask implements TaskInterface, CommandInterface
         $this->result = Result::success($this);
     }
 
+    /**
+     * Executes `git clone`
+     *
+     * @param $repo
+     * @param string $to
+     * @return $this
+     */
     public function cloneRepo($repo, $to = "")
     {
         $this->stackCommands[]= "clone $repo $to";
         return $this;
     }
 
+    /**
+     * Git commands in stack will stop if any of commands were unsuccessful
+     *
+     * @return $this
+     */
     public function stopOnFail()
     {
         $this->stopOnFail = true;
         return $this;
     }
 
+    /**
+     * Executes `git add` command with files to add pattern
+     *
+     * @param $pattern
+     * @return $this
+     */
     public function add($pattern)
     {
         $this->stackCommands[]= "add $pattern";
         return $this;
     }
 
+    /**
+     * Executes `git commit` command with a message
+     *
+     * @param $message
+     * @param string $options
+     * @return $this
+     */
     public function commit($message, $options = "")
     {
         $this->stackCommands[] = "commit -m '$message' $options";
         return $this;
     }
 
+    /**
+     * Executes `git pull` command.
+     *
+     * @param string $origin
+     * @param string $branch
+     * @return $this
+     */
     public function pull($origin = '', $branch = '')
     {
         $this->stackCommands[] = "pull $origin $branch";
         return $this;        
     }
 
+    /**
+     * Executes `git push` command
+     *
+     * @param string $origin
+     * @param string $branch
+     * @return $this
+     */
     public function push($origin = '', $branch = '')
     {
         $this->stackCommands[] = "push $origin $branch";
         return $this;
     }
 
+    /**
+     * Executes `git checkout` command
+     *
+     * @param $branch
+     * @return $this
+     */
     public function checkout($branch)
     {
         $this->stackCommands[] = "checkout $branch";